Check that everything works with ptracer and pt-run from /usr/local
It's inconvenient to do make install every time we want to run the test suite.
So I set it up to run the /usr/local tests only in the Github CI. To test that
the compilation works with /usr/local version of ptracer.h, clear the CPPFLAGS
variable when compiling the tests.